Update publishing-releases.rst
I have noticed that all examples use 'python3'. however, two last examples on this page use 'python. I propose to fix the typo. Also, it came to my attention that 'wheel' is not a standard package and needs to be installed manual I propose to add it to the instructions.